renamed output_details files to snakecase
