+++ src/osdag_gui/ui/components/output_details/b2b_cover_plate_welded.py
