+++ src/osdag_gui/ui/components/output_details/b2b_end_plate_sketch.py
