+++ src/osdag_gui/ui/components/output_details/b2c_end_plate.py
