+++ src/osdag_gui/ui/components/output_details/base_plate_hollow.py
