+++ src/osdag_gui/ui/components/output_details/bolt_pattern.py
