+++ src/osdag_gui/ui/components/output_details/c2c_end_plate.py
