+++ src/osdag_gui/ui/components/output_details/cleat_angle.py
