+++ src/osdag_gui/ui/components/output_details/end_plate.py
