+++ src/osdag_gui/ui/components/output_details/fin_plate_capacity.py
