+++ src/osdag_gui/ui/components/output_details/seated_angle_spacing.py
