+++ src/osdag_gui/ui/components/output_details/tension_bolted_spacing.py
